criteria_processor_OLD/requirements.txt
pandas>=2.2.3
pyyaml>=6.0.2
python-dotenv>=1.1.0
openai>=1.78.1
requests>=2.31.0 